adventofcode/2021/requirements.txt
advent-of-code-data==1.1.1
python-dotenv==0.19.2